+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.1.0
+### Fixes and new methods.
+- `GetRandomElement` methods renamed to `GetRandom`. 
+- `Except` and `Shuffled` methods now return list. 
+- Added `Shuffle` and `Swap` methods to `IList`. 
+- Added `AsRect` method to `Transform`. Added `Rotate` method to `Vector2`.
+
 # v1.0.6
 ### fixes and updates:
 - Added support for `Approx` and `WithRandomSign` methods to `double`.